ml/requirements.txt
flask
requests
python-dotenv
google-generativeai
scikit-learn
apscheduler
numpy
pandas
joblib
